refactor: rename component files to pascal case
